[FIX] *: stop handling record ids as 32-bit integers
Now that a new database numbers its records with int8, the places that still
treat an id as a 32-bit integer break. They fall in four groups.

Queries that cast an id back to a 32-bit integer, which raises "integer out of
range" as soon as the id does not fit: account splits parent_path to join the
ancestors of an account and builds the VALUES list of its recurring entries,
project reads the stage id from the keys of duration_tracking, website compares
website_visitor.partner_id with its access_token, website_sale splits the
parent_path of a category, cloud_storage_migration reads an attachment id from
ir.config_parameter, and l10n_fr_pos_cert takes the company id from the first
part of parent_path.

A query that compared arrays of ids: mail looked for an existing chat by
comparing the aggregated partner ids of a channel with the ones it was given.
PostgreSQL has no equality operator between arrays whose element types differ,
so the query failed outright; casting both sides makes it work on either width.

Queries that turned an id into a boolean: account tested BOOL(pay.id) to know
whether a payment exists, and its journal dashboard did the same to know whether
a journal has entries. PostgreSQL defines a cast from int4 to bool but none
from int8, so the query failed on any database with 64-bit ids, whatever their
value. Testing the column against NULL says what the code means and works
whatever the width is.

Columns created outside of the ORM with a hardcoded width, to avoid computing a
stored field on a large table while installing: the pre-init hook of
l10n_fr_pdp, and the fields that store the id of a record without declaring a
relation to it. The latter now use the `bigint` argument of Integer:

- the res_id of ir.actions.act_window, ir.filters, ir.embedded.actions and of
  the portal, sms, snailmail, payment, privacy_lookup and web models
- res.device.user_id, base.partner.merge.min_id, purchase's matching_id,
  event_crm's processed_registration_id, im_livechat's script answer id,
  mail.alias' thread ids and hr.expense's former sheet
- the ids deliberately kept as a number so they survive the deletion of the
  record they point to: ir.logging's create_uid and write_uid, whose foreign
  key the module drops on purpose, res.users.deletion.user_id_int, the sms and
  mail ids of mail.notification and mailing.trace, and the old and new value of
  mail.tracking.value, which hold the id of a tracked many2one next to its
  display name
- the counters that grow once per record, which are exposed to the same
  overflow as an id: the number_next of ir.sequence and of its date ranges,
  mirroring a PostgreSQL sequence that is bigint, and the closing number of
  l10n_fr_pos_cert

Counters bounded by the activity on a single record are left as they are:
res.partner.customer_rank, res.users.karma, the bounce counter of a contact or
the print counter of a POS order grow with what happens to one record, not with
the size of the database. Integer fields holding an id that belongs to a
third-party system are left alone as well, and so are the advisory lock keys
read from pg_locks, which pg_advisory_lock defines as 32-bit.

+++ addons/account/models/account_account.py
@@ -422,7 +422,7 @@ class AccountAccount(models.Model):
                            STRING_AGG(%(ancestors_names)s, ' / ' ORDER BY ord) AS name_path
                       FROM account_account child
         CROSS JOIN LATERAL UNNEST(STRING_TO_ARRAY(RTRIM(child.parent_path, '/'), '/')) WITH ORDINALITY AS t(id, ord)
-                      JOIN account_account ancestor ON ancestor.id = t.id::int
+                      JOIN account_account ancestor ON ancestor.id = t.id::bigint
                   GROUP BY child.id
                 )""",
                 ancestors_ids=ancestor_table.id,
